Starting inout port splitting
