+++ tests/unit/split_inout_port.cpp
@@ -1,0 +1,12 @@
+#include "coreir.h"
+//#include "coreir/passes/transform/split_inouts.h"
+
+using namespace CoreIR;
+using namespace std;
+
+int main() {
+  Context* c = newContext();
+
+  deleteContext(c);
+  assert(false);
+}
